requirements.txt
ai21==2.4.1
ai21-tokenizer==0.9.1
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.4
boto3==1.34.135
botocore==1.34.135
bs4==0.0.2
build==1.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
constantly==23.10.4
cppyy==2.4.2
cppyy-backend==1.14.10
cppyy-cling==6.27.1
CPyCppyy==1.12.12
cron-descriptor==1.4.3
cryptography==42.0.5
curlify==2.2.1
daphne==4.1.2
dataclasses-json==0.6.4
datetime-truncate==1.1.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
distro-info==1.1+ubuntu0.1
Django==5.0.2
django-admin-charts==1.3.1
django-admin-logs==1.2.0
django-allauth==0.61.1
django-bower==5.2.0
django-channels==0.7.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-debug-toolbar==4.3.0
django-eventstream==5.1.0
django-filter==24.2
django-grip==3.5.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.10.1
django-redis==5.4.0
django-rest-passwordreset==1.4.0
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework==3.14.0
drf-nested-routers==0.94.1
elastic-transport==8.13.1
elasticsearch==8.14.0
exceptiongroup==1.2.1
executing==2.0.1
facebook_business==19.0.2
faiss-cpu==1.8.0
fastapi==0.110.3
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
google-ads==23.1.0
google-ai-generativelanguage==0.6.6
google-analytics==0.0.0
google-analytics-admin==0.22.7
google-analytics-data==0.18.8
google-api-core==2.18.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
gripcontrol==4.2.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.25.0
iso8601==2.1.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jq==1.7.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==23.5.0
kombu==5.3.7
kubernetes==29.0.0
langchain==0.2.6
langchain-ai21==0.1.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-elasticsearch==0.2.2
langchain-experimental==0.0.62
langchain-google-genai==1.0.7
langchain-openai==0.1.11
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langchainhub==0.1.15
langsmith==0.1.82
lark==1.1.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
linkedin-api-client==0.3.0
mailchimp-marketing==3.0.80
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.8.1
ns3==3.40
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.35.7
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pagekite==1.5.2.200603
parso==0.8.4
pexpect==4.9.0
pgvector==0.2.5
pillow==10.2.0
pinecone-client==3.2.2
pinecone-text==0.9.0
platformdirs==4.2.2
posthog==3.5.0
prompt_toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pubcontrol==3.5.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==23.12.11
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
python-apt==2.4.0+ubuntu2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rank-bm25==0.2.2
redis==5.0.7
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.1
sentencepiece==0.2.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sob==1.50.8
SocksipyChain==2.1.2
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
systemd-python==234
tenacity==8.3.0
text-unidecode==1.3
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
Twisted==24.3.0
twitter-ads==11.0.0
txaio==23.1.1
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
virtualenv==20.26.1
wadllib==1.3.6
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wget==3.2
whitenoise==6.6.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
zope.interface==6.4.post2
